clean up for readability
