+++ lib/apis.js
@@ -1,0 +1,5 @@
+const path = require('path')
+const contentDir = path.join(__dirname, '../content')
+
+module.exports = require(path.join(contentDir, 'en/api/electron-api.json'))
+  .sort((a, b) => a.name.toLowerCase().localeCompare(b.name.toLowerCase()))
